new_apps/requirements.txt
requests
python-dotenv
google-generativeai
google-genai
streamlit
agno
google-search-results